@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ValidatorFn, ValidationErrors, AbstractControl, Validators } from '@angular/forms';
+import { ValidatorFn, ValidationErrors, AbstractControl, Validators, FormControl } from '@angular/forms';
 import { DataDefinitionService } from '@service/data-definition/data-definition.service';
 import { SessionStorageService } from '@service/storage/session-storage.service';
 
@@ -29,12 +29,12 @@ export class LocalValidators {
        * Los campos pueden crearse dinamicamente, por lo que debe verificarse si se encuentran definidos
        */
 
-      const cuil = control.get("cuil").value;
-      const numeroDocumento = control.get("numero_documento").value;
+      const cuil = (control.get("cuil") as FormControl).value;
+      const numeroDocumento = (control.get("numero_documento") as FormControl).value;
       
       if(cuil && numeroDocumento){
-        const cuil_ = control.get("cuil").value.substring(2,10).replace(/^0+/, '');          
-        const numeroDocumento_ = control.get("numero_documento").value.replace(/^0+/, '');
+        const cuil_ = (control.get("cuil") as FormControl).value.substring(2,10).replace(/^0+/, '');          
+        const numeroDocumento_ = (control.get("numero_documento") as FormControl).value.replace(/^0+/, '');
         if(cuil_ != numeroDocumento_) {
           return {cuilDni:true}
         }
